requirements.txt
flask
python-dotenv
google-genai
gspread
google-auth
